requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
python-dotenv
pydantic
requests
beautifulsoup4
PyPDF2
arxiv
newspaper3k
duckduckgo_search
wikipedia
